+++ stockhealth/utilities/calendar.py
@@ -13,7 +13,7 @@ from stockhealth.model import _NUMBER_OF_CALENDAR_DAYS_PER_YEAR as _NCD
 _NUMBER_OF_SECONDS_PER_DAY: float = 86400
 
 
-def dt(now: date or np.float_, reference: date or np.float_) -> np.float:
+def dt(now: date or float_, reference: date or float_) -> float:
     """Give the date delta."""
     if type(now) is date:
         days = (now - reference).days + \
